refactor: 패키지 변경 및 코드 수정
